refactor code to apply SOLID principles
