+++ book.rb
@@ -12,5 +12,4 @@ class Book
     @rentals.push(rental)
     rental.book = self
   end
-  
 end
