--- input_reader.rb
@@ -1,5 +1,0 @@
-class InputReader
-  def read_input
-    gets.chomp
-  end
-end
