+++ person.rb
@@ -1,20 +1,15 @@
 require_relative 'nameable'
 
 class Person < Nameable
-  attr_accessor :name, :age, :parent_permission
+  attr_accessor :name, :age
   attr_reader :id, :rentals
 
-  def initialize(age, name = 'unknown', parent_permission: true)
+  def initialize(age, name = 'unknown')
     super()
     @id = Random.rand(1..1000)
     @age = age
     @name = name
-    @parent_permission = parent_permission
     @rentals = []
-  end
-
-  def can_use_services?
-    is_of_age? || @parent_permission
   end
 
   def correct_name
@@ -24,11 +19,5 @@ class Person < Nameable
   def add_rental(rental)
     @rentals.push(rental)
     rental.book = self
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
   end
 end
