+++ teacher.rb
@@ -3,8 +3,8 @@ require_relative 'person'
 class Teacher < Person
   attr_accessor :specialization
 
-  def initialize(age, specialization, name = 'unknown', parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
+  def initialize(age, specialization, name)
+    super(age, name)
     @specialization = specialization
   end
